tests: reorganize fixtures into separate modules
